dreambot/pyproject.toml
[project]
name = "dreambot"
version = "0.1.0"
description = "DreamBot — NDX/SPX/QQQ/SPY Options Scalping & Day-Trading System"
authors = [{ name = "DreamBot Team" }]
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.9",
  "websockets>=12.0",
  "PyYAML>=6.0",
  "redis>=5.0",
  "duckdb>=0.10",
  "streamlit>=1.32",
  "pandas>=2.2",
  "numpy>=1.26",
  "scipy>=1.10",
  "scikit-learn>=1.4"
]

[project.optional-dependencies]
test = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "freezegun>=1.4",
  "responses>=0.25"
]

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["services*"]
